Added Blinn-Phone light with gamma correction + bug fixes

+++ src/Core/Engine/Events/WindowResizeEvent.hpp
@@ -7,8 +7,8 @@ class WindowResizeEvent : public Event
 {
 public:
     GLFWwindow* Window;
-    int Width, Height;
+    double Width, Height;
 
-    WindowResizeEvent(GLFWwindow* window, int width, int height) : Window(window), Width(width), Height(height) {}
+    WindowResizeEvent(GLFWwindow* window, double width, double height) : Window(window), Width(width), Height(height) {}
     EVENT_TYPE(WindowResizeEvent);
 };
